Add script to find high churn rate files in git
Script that allows you to find files that change often.

Use commands like:
- git churn --since='3 months ago'
- git churn dao service
- git churn domain core
- git churn

The script will then output the number of times files has changed,
giving you an indication of which files that changes the most often and
thus is probably the most coupled (a.k.a PITAs).

In any case, the higher order files are the ones that is a target for
refactoring.